refactor(uptime): Extract assertion feature flags into useUptimeAssertionFeatures hook
