+++ static/app/views/detectors/components/forms/uptime/verification.tsx
@@ -1,14 +1,14 @@
 import {Container} from 'sentry/components/workflowEngine/ui/container';
 import Section from 'sentry/components/workflowEngine/ui/section';
 import {t} from 'sentry/locale';
-import useOrganization from 'sentry/utils/useOrganization';
 import {UptimeAssertionsField} from 'sentry/views/alerts/rules/uptime/assertions/field';
+import {useUptimeAssertionFeatures} from 'sentry/views/alerts/rules/uptime/useUptimeAssertionFeatures';
 import {UptimeSectionGrid} from 'sentry/views/detectors/components/forms/uptime/styles';
 
 export function UptimeDetectorVerificationSection() {
-  const org = useOrganization();
+  const {hasRuntimeAssertions} = useUptimeAssertionFeatures();
 
-  if (!org.features.includes('uptime-runtime-assertions')) {
+  if (!hasRuntimeAssertions) {
     return null;
   }
